Fix tests and update examples.
